+ Added docs for ApiRenderItem and ApiRenderClient * Moved some render methods for blocks to their own class
